refactor(react): migrate issue + experimental-issue stores to useAppStore
Add listIssues to the Zustand issue slice (alongside the existing
fetchIssueByName) and relocate the issue-related helpers out of the Pinia
layer: experimentalCreateIssueByPlan + buildIssueFilter into
react/stores/app/issue.ts, candidatesOfApprovalStepV1 into an issue-detail
util, and refreshIssueList/useRefreshIssueList into a Vue-free react helper.
Migrate all React consumers and delete the now-dead Pinia issue and
experimental-issue modules + their barrel exports.
