+++ frontend/src/store/modules/v1/index.ts
@@ -5,11 +5,9 @@ export * from "./database";
 export * from "./databaseCatalog";
 export * from "./dbSchema";
 export * from "./environment";
-export * from "./experimental-issue";
 export * from "./group";
 export * from "./instance";
 export * from "./instanceResource";
-export * from "./issue";
 export * from "./permission";
 export * from "./policy";
 export * from "./project";
